requirement.txt
Flask
Flask-Cors
Pillow
tensorflow
numpy
